requirements.txt
robotframework==7.0.1
robotframework-seleniumlibrary==6.3.0
selenium==4.16.0
webdriver-manager==4.0.1
